requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.16
strawberry-graphql>=0.288
strawberry-graphql-django>=0.72
django-htmx>=1.27
django-tailwind>=4.4
